test: `requires_latch_api` marker (#2)
## Summary

This PR adds a pytest marker, `requires_latch_api`, to annotate and
xfail unit tests which require online access to Latch.

Tests will require access to the Fulcrum Genomics workspace; I've
hard-coded this into the marker and provided documentation in the
CONTRIBUTING and a corresponding warning in the marker if a user is
running tests locally but in another workspace.

See #3 for an example in practice:


https://github.com/fulcrumgenomics/fglatch/blob/f16804129ae94b5c72d5d645736f9963ede29e4e/tests/client/test_executions.py#L9-L10